refactor(ui): split ipc.ts into transport seam + command-family modules (#1200)
TD-054: ipc.ts (~740 LOC) combined transport selection, event subscription,
and every command wrapper. Split into a stable barrel:
- ipc/transport.ts — the single transport adapter (command, onEvent,
  onReconnect, disposeTransport, isTauri, WebSocket internals).
- ipc/{game,save,setup,provider,screenshot}.ts — command families importing
  the transport seam.
ipc.ts re-exports everything so all '$lib/ipc' / './ipc' import paths and
signatures are unchanged. svelte-check + 700 vitest tests pass; eslint clean.
